requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
cmake==3.25.2
ffmpeg-python==0.2.0
filelock==3.9.1
Flask==2.2.3
future==0.18.3
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
lit==15.0.7
llvmlite==0.39.1
MarkupSafe==2.1.2
more-itertools==9.1.0
mpmath==1.3.0
networkx==3.0
numba==0.56.4
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
openai-whisper @ git+https://github.com/openai/whisper.git@79c43e48592bd5cd9d893236a617d2ddeee8d878
Pillow==9.4.0
pydub==0.25.1
regex==2022.10.31
requests==2.28.2
sympy==1.11.1
tiktoken==0.3.1
torch==1.13.1
torchaudio==0.13.1
torchvision==0.14.1
tqdm==4.65.0
triton==2.0.0
typing_extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.2.3
zipp==3.15.0
